Export settings #51

+++ app/components/VideoExport.js
@@ -18,6 +18,7 @@ import {
 } from '../utils/utils';
 import Airing from '../utils/Airing';
 import { CHECKBOX_OFF, CHECKBOX_ON } from './Checkbox';
+import getConfig from '../utils/config';
 
 type Props = {
   airingList: Array<Airing>,
@@ -28,7 +29,12 @@ type Props = {
   bulkRemExportRecord: (Array<ExportRecordType>) => void
 };
 
-type State = { exportState: number, atOnce: number, deleteOnFinish: number };
+type State = {
+  exportState: number,
+  atOnce: number,
+  deleteOnFinish: number,
+  actionOnDuplicate: string
+};
 
 const VideoExport = (WrappedComponent: any) => {
   // $FlowFixMe wtf typing
@@ -46,7 +52,8 @@ const VideoExport = (WrappedComponent: any) => {
       this.state = {
         exportState: EXP_WAITING,
         atOnce: 1,
-        deleteOnFinish: CHECKBOX_OFF
+        deleteOnFinish: CHECKBOX_OFF,
+        actionOnDuplicate: getConfig().actionOnDuplicate
       };
 
       this.shouldCancel = false;
@@ -66,9 +73,11 @@ const VideoExport = (WrappedComponent: any) => {
 
     processVideo = async () => {
       const { exportList } = this.props;
-      const { exportState, atOnce } = this.state;
+      const { atOnce, actionOnDuplicate } = this.state;
 
-      if (exportState === EXP_DONE) return;
+      global.EXPORTING = true;
+
+      // if (exportState === EXP_DONE) return;
       await this.setState({ exportState: EXP_WORKING });
 
       const actions = [];
@@ -76,7 +85,10 @@ const VideoExport = (WrappedComponent: any) => {
       exportList.forEach(rec => {
         actions.push(() => {
           if (this.shouldCancel === false)
-            return rec.airing.processVideo(this.updateProgress);
+            return rec.airing.processVideo(
+              actionOnDuplicate,
+              this.updateProgress
+            );
         });
       });
 
@@ -89,6 +101,7 @@ const VideoExport = (WrappedComponent: any) => {
       } else {
         this.setState({ exportState: EXP_DONE });
       }
+      global.EXPORTING = false;
     };
 
     updateProgress = (airingId: number, progress: Object) => {
@@ -108,7 +121,10 @@ const VideoExport = (WrappedComponent: any) => {
 
       if (progress.finished) {
         if (deleteOnFinish === CHECKBOX_ON) {
-          airing.delete();
+          const status = airing.isExportValid();
+          if (status.valid) {
+            airing.delete();
+          }
         }
         record.state = EXP_DONE;
         record.progress = {
@@ -178,20 +194,31 @@ const VideoExport = (WrappedComponent: any) => {
       });
     };
 
+    setActionOnDuplicate = (event: SyntheticEvent<HTMLInputElement>) => {
+      this.setState({ actionOnDuplicate: event.currentTarget.value });
+    };
+
     render() {
-      const { exportState, deleteOnFinish, atOnce } = this.state;
+      const {
+        exportState,
+        actionOnDuplicate,
+        deleteOnFinish,
+        atOnce
+      } = this.state;
       /* eslint-disable react/jsx-props-no-spreading */
       // $FlowFixMe
       return (
         <WrappedComponent
           {...this.props}
           exportState={exportState}
+          actionOnDuplicate={actionOnDuplicate}
+          setActionOnDuplicate={this.setActionOnDuplicate}
           deleteOnFinish={deleteOnFinish}
+          toggleDOF={this.toggleDOF}
           atOnce={atOnce}
           atOnceChange={this.atOnceChange}
           processVideo={this.processVideo}
           cancelProcess={this.cancelProcess}
-          toggleDOF={this.toggleDOF}
         />
       );
     }
